Implemented Exception Handling

+++ src/main/java/com/shruthan/jobportalbackend/service/ApplicationService.java
@@ -6,6 +6,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.shruthan.jobportalbackend.exception.InvalidInputException;
+import com.shruthan.jobportalbackend.exception.ResourceNotFoundException;
 import com.shruthan.jobportalbackend.model.Application;
 import com.shruthan.jobportalbackend.repository.ApplicationRepository;
 
@@ -28,14 +30,33 @@ public class ApplicationService {
 	}
 
 	public Application getApplicationById(String id) {
-		return applicationRepository.findById(id).orElse(null);
+		Application application = applicationRepository.findById(id).orElse(null);
+		
+		if (application == null) {
+			throw new ResourceNotFoundException("Job Application Not Found");
+		}
+		
+		return application;
 	}
 
 	public void updateApplicationById(Application application, String id) {
-		applicationRepository.save(application);
+		
+		if (!application.getId().equals(id)) {
+			throw new InvalidInputException("Job Application Id and input object Id doesn't match");
+		}
+		
+		if (applicationRepository.findById(id).isEmpty()) {
+			throw new ResourceNotFoundException("Job Application Not Found");
+		}
+		
+		applicationRepository.save(application);	
 	}
 
 	public void deleteApplicationById(String id) {
+		
+		if (applicationRepository.findById(id).isEmpty()) {
+			throw new ResourceNotFoundException("Job Application Not Found");
+		}
 		applicationRepository.deleteById(id);
 	}
